spaced: use proton clang

+++ build.sh
@@ -3,40 +3,43 @@
 function compile() 
 {
 
+source ~/.bashrc && source ~/.profile
+export LC_ALL=C && export USE_CCACHE=1
+ccache -M 100G
 export ARCH=arm64
-export KBUILD_BUILD_HOST="flow-eng-labs"
-export KBUILD_BUILD_USER="rk134"
-export USE_CCACHE=1
-ccache -M 5G
-git clone --depth=1 https://github.com/sarthakroy2002/android_prebuilts_clang_host_linux-x86_clang-6443078 $HOME/toolchains/clang
-git clone --depth=1 https://github.com/LineageOS/android_prebuilts_gcc_linux-x86_aarch64_aarch64-linux-android-4.9 $HOME/toolchains/los-4.9-64
-git clone --depth=1 https://github.com/LineageOS/android_prebuilts_gcc_linux-x86_arm_arm-linux-androideabi-4.9 $HOME/toolchains/los-4.9-32
+export KBUILD_BUILD_HOST=COSMOS
+export KBUILD_BUILD_USER="INFIX"
+git clone --depth=1 https://github.com/kdrag0n/proton-clang clang
+git clone --depth=1 https://github.com/sarthakroy2002/prebuilts_gcc_linux-x86_aarch64_aarch64-linaro-7 los-4.9-64
+git clone --depth=1 https://github.com/sarthakroy2002/linaro_arm-linux-gnueabihf-7.5 los-4.9-32
 
-[ -d "out" ] && rm -rf out || mkdir -p out
+[ -d "out" ] && rm -rf AnyKernel && rm -rf out || mkdir -p out
 
 make O=out ARCH=arm64 spaced_defconfig
 
-PATH="$HOME/toolchains//clang/bin:${PATH}:$HOME/toolchains//los-4.9-32/bin:${PATH}:$HOME/toolchains//los-4.9-64/bin:${PATH}" \
+PATH="${PWD}/clang/bin:${PATH}:${PWD}/los-4.9-32/bin:${PATH}:${PWD}/los-4.9-64/bin:${PATH}" \
 make -j$(nproc --all) O=out \
                       ARCH=arm64 \
-                      CC="ccache clang" \
+                      CC="clang" \
                       CLANG_TRIPLE=aarch64-linux-gnu- \
-                      CROSS_COMPILE="$HOME/toolchains/los-4.9-64/bin/aarch64-linux-android-" \
-                      CROSS_COMPILE_ARM32="$HOME/toolchains/los-4.9-32/bin/arm-linux-androideabi-" \
-                      CONFIG_NO_ERROR_ON_MISMATCH=y
+                      CROSS_COMPILE="${PWD}/los-4.9-64/bin/aarch64-linux-gnu-" \
+                      CROSS_COMPILE_ARM32="${PWD}/los-4.9-32/bin/arm-linux-gnueabihf-" \
+                      LD=ld.lld \
+                      AS=llvm-as \
+		              AR=llvm-ar \
+			          NM=llvm-nm \
+			          OBJCOPY=llvm-objcopy \
+                      CONFIG_SECTION_MISMATCH_WARN_ONLY=y
 }
 
 function zupload()
 {
-rm -rf AnyKernel
-git clone --depth=1 https://github.com/rk134/AnyKernel3.git AnyKernel -b flow
-cp out/arch/arm64/boot/Image.gz AnyKernel
-cp out/arch/arm64/boot/dtb AnyKernel
-cp out/arch/arm64/boot/dtbo.img AnyKernel
+git clone --depth=1 https://github.com/HELLINFIX/AnyKernel3.git AnyKernel
+cp out/arch/arm64/boot/Image.gz-dtb AnyKernel
 cd AnyKernel
-zip -r9 flowKernel-OSS-S-source-spaced.zip *
+zip -r9 Cosmos-kernel-1.0-spaced.zip *
 curl -sL https://git.io/file-transfer | sh
-./transfer wet flowKernel-OSS-S-source-spaced.zip
+./transfer wet Cosmos-kernel-1.0-spaced.zip
 }
 
 compile
